Move some type variable code into a separate file
