+++ clang/lib/CConv/ConstraintBuilder.cpp
@@ -14,6 +14,7 @@
 #include "clang/CConv/ConstraintResolver.h"
 #include "clang/CConv/ArrayBoundsInferenceConsumer.h"
 #include "clang/CConv/CCGlobalOptions.h"
+#include "clang/CConv/TypeVariableAnalysis.h"
 
 using namespace llvm;
 using namespace clang;
@@ -599,63 +600,4 @@ void ConstraintBuilderConsumer::HandleTranslationUnit(ASTContext &C) {
 
   Info.exitCompilationUnit();
   return;
-}
-
-std::set<ConstraintVariable *> &TypeVariableEntry::getConstraintVariables() {
-  assert("Accessing ConstraintVariable set for inconsistent Type Variable." &&
-      IsConsistent);
-  return ArgConsVars;
-}
-
-void TypeVariableEntry::insertConstraintVariables
-    (set<ConstraintVariable *> &CVs) {
-  assert("Accessing ConstraintVariable set for inconsistent Type Variable." &&
-      IsConsistent);
-  ArgConsVars.insert(CVs.begin(), CVs.end());
-}
-
-void TypeVariableEntry::setTypeParamConsVar(ConstraintVariable *CV) {
-  assert("Accessing constraint variable for inconsistent Type Variable." &&
-      IsConsistent);
-  assert("Setting constraint variable to null" && CV != nullptr);
-  assert("Changing already set constraint variable" &&
-      TypeParamConsVar == nullptr);
-  TypeParamConsVar = CV;
-}
-
-void TypeVariableEntry::updateEntry(QualType Ty,
-                                    std::set<ConstraintVariable *> &CVs) {
-  QualType PointeeType = Ty->getPointeeType();
-  if (isTypeAnonymous(PointeeType)) {
-    // We'll need a name to provide the type arguments during rewriting, so
-    // no anonymous things here.
-    IsConsistent = false;
-  } else if (IsConsistent && getType() != Ty) {
-    // If it has previously been instantiated as a different type, its use
-    // is not consistent.
-    IsConsistent = false;
-  } else if (IsConsistent) {
-    // Type variable has been encountered before with the same type. Insert
-    // new constraint variables.
-    insertConstraintVariables(CVs);
-  }
-  // If none of the above branches are hit, then this was already inconsistent,
-  // so there's no need to update anything.
-}
-
-ConstraintVariable *TypeVariableEntry::getTypeParamConsVar() {
-  assert("Accessing constraint variable for inconsistent Type Variable." &&
-      IsConsistent);
-  assert("Accessing null constraint variable" && TypeParamConsVar != nullptr);
-  return TypeParamConsVar;
-}
-
-QualType TypeVariableEntry::getType() {
-  assert("Accessing Type for inconsistent Type Variable." &&
-      IsConsistent);
-  return TyVarType;
-}
-
-bool TypeVariableEntry::getIsConsistent() {
-  return IsConsistent;
 }
